#74 리팩터링; enums 모듈 생성;
